+++ packages/inngest/src/components/connect/index.ts
@@ -1,6 +1,15 @@
+import { MessageHandler } from "./messages";
+import type {
+  ConnectApp,
+  ConnectHandlerOptions,
+  ConnectionState,
+  DEFAULT_SHUTDOWN_SIGNALS,
+  WorkerConnection,
+} from "./types";
+
 // Export types for convenience
 export {
-  DEFAULT_SHUTDOWN_SIGNALS,
+  type DEFAULT_SHUTDOWN_SIGNALS,
   type ConnectApp,
   type ConnectHandlerOptions,
   ConnectionState,
@@ -10,7 +19,7 @@ export {
 export const connect = async (
   options: ConnectHandlerOptions
 ): Promise<WorkerConnection> => {
-  const conn = new WebSocketWorkerConnection(options);
+  const conn = new MessageHandler(options);
 
   await conn.start();
